feat: Limit NewEmitter to one instance by default
Change the API of NewEmitter to return an error when there is more than
one instance running already.

Add an AllowMultipleInstances in case we have a need for running
multiple instances at some point for some reason.

This paves the path for reintroducing the old Start() API and is
probably also a good idea in general.